requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
autobahn==22.12.1
Automat==22.10.0
cffi==1.15.1
channels==4.0.0
channels-redis==4.0.0
constantly==15.1.0
cryptography==38.0.4
daphne==4.0.0
Django==4.1.3
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-api-key==2.2.0
djangorestframework-simplejwt==5.2.2
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
iniconfig==1.1.1
msgpack==1.0.4
packaging==22.0
Pillow==9.3.0
pluggy==1.0.0
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.6.0
pyOpenSSL==22.1.0
pytest==7.2.0
pytest-asyncio==0.20.1
pytest-django==4.5.2
pytz==2022.7
redis==4.4.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.3
Twisted==22.10.0
txaio==22.2.1
typing_extensions==4.4.0
zope.interface==5.5.2
